refactor(Draw): 당첨 번호 관련 기능을 Draw 클래스로 분리
