+++ src/Lotto.js
@@ -1,5 +1,3 @@
-const { SEPARATOR } = require('./settings');
-
 class Lotto {
   #numbers;
 
@@ -14,9 +12,6 @@ class Lotto {
     }
   }
 
-  static changeWinningNumbersType(input) {
-    return input.split(SEPARATOR).map((winningNumber) => Number(winningNumber));
-  }
   // TODO: 추가 기능 구현
 }
